# docs/swagger.yaml
basePath: /
definitions:
  model.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: integer
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        description: Возвращаем uuid.UUID
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для агрегации данных об онлайн-подписках пользователей
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
      summary: Получить список всех подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
      summary: Создать новую подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: ID подписки
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
      summary: Обновить данные подписки
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      parameters:
      - description: ID пользователя (UUID)
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Начало периода (MM-YYYY)
        in: query
        name: start_period
        required: true
        type: string
      - description: Конец периода (MM-YYYY)
        in: query
        name: end_period
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              total:
                type: integer
            type: object
      summary: Подсчёт суммарной стоимости подписок за период
      tags:
      - subscriptions
swagger: "2.0"